+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+v2.3.0
+------
+
+- Added group restriction check #54
+- Added group update observer to amend grades when group membership changes #37
+- Made WebPA calculator a dependency
+- Added download all submissions
+- Changed icon
+- Added first and last names to export #48
+- Added student ID to export #57
+- Fixed bug in calculation when self grade turned off #58
+
+v2.2.0
+------
+
+- Fixed single calculator plugin error #55 #56
+- Added support for v3.10
+
 v2.1.0
 ------